+++ app/lib/question_answer/post_action_type_extension.rb
@@ -1,0 +1,7 @@
+module QuestionAnswer
+  module PostActionTypeExtension
+    def public_types
+      @public_types ||= super.except(:vote)
+    end
+  end
+end
