+++ app/lib/question_answer/post_creator_extension.rb
@@ -1,0 +1,8 @@
+module QuestionAnswer
+  module PostCreatorExtension
+    def valid?
+      guardian.post_opts = @opts
+      super
+    end
+  end
+end
