+++ app/lib/question_answer/topic_list_item_serializer_extension.rb
@@ -1,0 +1,23 @@
+module QuestionAnswer
+  module TopicListItemSerializerExtension
+    attributes :qa_enabled,
+               :answer_count
+
+    def qa_enabled
+      true
+    end
+
+    def include_qa_enabled?
+      Topic.qa_enabled object
+    end
+
+    def answer_count
+      object.answer_count
+    end
+
+    def include_answer_count?
+      include_qa_enabled?
+    end
+  end
+
+end
