+++ app/lib/question_answer/voter.rb
@@ -1,0 +1,1 @@
+Voter = Struct.new(:user)
