+++ lib/qa.rb
@@ -1,1 +1,1 @@
-require_relative 'question_answer/engine'
+# require_relative 'question_answer/engine'
